+++ src/API/Common/Time/ClockFactoryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\SDK\Common\Time;
+namespace OpenTelemetry\API\Common\Time;
 
 interface ClockFactoryInterface
 {
