+++ src/SDK/Common/Time/Util.php
@@ -4,29 +4,31 @@ declare(strict_types=1);
 
 namespace OpenTelemetry\SDK\Common\Time;
 
+use OpenTelemetry\API\Common\Time\Util as API;
+
+/**
+ * @deprecated Use OpenTelemetry\API\Common\Time\Util
+ * @codeCoverageIgnore
+ */
 class Util
 {
-    /** @psalm-pure */
     public static function nanosToMicros(int $nanoseconds): int
     {
-        return intdiv($nanoseconds, ClockInterface::NANOS_PER_MICROSECOND);
+        return API::nanosToMicros($nanoseconds);
     }
 
-    /** @psalm-pure */
     public static function nanosToMillis(int $nanoseconds): int
     {
-        return intdiv($nanoseconds, ClockInterface::NANOS_PER_MILLISECOND);
+        return API::nanosToMillis($nanoseconds);
     }
 
-    /** @psalm-pure */
     public static function secondsToNanos(int $seconds): int
     {
-        return $seconds * ClockInterface::NANOS_PER_SECOND;
+        return API::secondsToNanos($seconds);
     }
 
-    /** @psalm-pure */
     public static function millisToNanos(int $milliSeconds): int
     {
-        return $milliSeconds * ClockInterface::NANOS_PER_MILLISECOND;
+        return API::millisToNanos($milliSeconds);
     }
 }
